requirements.txt
Flask
python-docx
pdfminer.six
